service-2-code-analysis/requirements.txt
fastapi==0.115.11
flasgger==0.9.7.1
Flask==3.1.0
flask_restful==0.3.10
GitPython==3.1.44
GitPython==3.1.44
pyairtable==3.0.2
pydantic==2.10.6
Requests==2.32.3
uvicorn==0.34.0
urllib3==2.3.0
regex==2024.11.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==1.12.3
argh==0.26.2